backend/requirements.txt
fastapi
uvicorn
google-generativeai
PyGithub
langgraph
python-dotenv